+++ tests/conftest.py
@@ -100,7 +100,7 @@ def setup_catalog(catalog_content, rmock):
 @pytest.fixture
 def produce_mock(mocker):
     mocker.patch("udata_hydra.crawl.send", dummy())
-    mocker.patch("udata_hydra.datalake_service.send", dummy())
+    mocker.patch("udata_hydra.analysis.send", dummy())
 
 
 @pytest.fixture
